dtype

+++ mantidimaging/core/io/saver.py
@@ -216,7 +216,7 @@ def _nexus_save(nexus_file: h5py.File, dataset: StrictDataset, sample_name: str)
 
     # instrument data
     combined_data_shape = (sum([len(arr) for arr in dataset.nexus_arrays]), ) + dataset.nexus_arrays[0].shape[1:]
-    detector.create_dataset("data", shape=combined_data_shape)
+    detector.create_dataset("data", shape=combined_data_shape, dtype="uint16")
     index = 0
     for arr in dataset.nexus_arrays:
         detector["data"][index:index + arr.shape[0]] = arr.astype("uint16")
